Add theme toggle and dynamic toast theme
